feat/DownloadFile: Download File Direct

+++ database/migrations/2025_10_13_192831_add_is_superuser_to_users_table.php
@@ -11,7 +11,9 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->boolean('is_superuser')->default(false)->after('password');
+            if (!Schema::hasColumn('users', 'is_superuser')) {
+                $table->boolean('is_superuser')->default(false);
+            }
         });
     }
 
@@ -21,8 +23,9 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_superuser');
+            if (Schema::hasColumn('users', 'is_superuser')) {
+                $table->dropColumn('is_superuser');
+            }
         });
     }
-
 };
